Move listen to a better file
This file is about listening, not sending notifications.

Also move the test to a matching file.
